revert "7536 use head for snap sync (#7718)" (#7767)
This reverts commit a7e1f6ace0167bbaff89f00f7a0e94747160a06b.
