+++ besu/src/main/java/org/hyperledger/besu/cli/ConfigurationOverviewBuilder.java
@@ -58,7 +58,6 @@ public class ConfigurationOverviewBuilder {
   private Integer trieLogsPruningWindowSize = null;
   private boolean isSnapServerEnabled = false;
   private boolean isSnapSyncBftEnabled = false;
-  private boolean isSnapSyncToHeadEnabled = true;
   private TransactionPoolConfiguration.Implementation txPoolImplementation;
   private EvmConfiguration.WorldUpdaterMode worldStateUpdateMode;
   private Map<String, String> environment;
@@ -259,18 +258,6 @@ public class ConfigurationOverviewBuilder {
   }
 
   /**
-   * Sets snap sync to head enabled/disabled
-   *
-   * @param snapSyncToHeadEnabled bool to indicate if snap sync to head is enabled
-   * @return the builder
-   */
-  public ConfigurationOverviewBuilder setSnapSyncToHeadEnabled(
-      final boolean snapSyncToHeadEnabled) {
-    isSnapSyncToHeadEnabled = snapSyncToHeadEnabled;
-    return this;
-  }
-
-  /**
    * Sets trie logs pruning window size
    *
    * @param size the max number of blocks to load and prune trie logs for at startup
@@ -400,10 +387,6 @@ public class ConfigurationOverviewBuilder {
 
     if (isSnapSyncBftEnabled) {
       lines.add("Experimental Snap Sync for BFT enabled");
-    }
-
-    if (isSnapSyncToHeadEnabled) {
-      lines.add("Snap Sync to Head enabled");
     }
 
     if (isBonsaiLimitTrieLogsEnabled) {
