+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/snapsync/SnapSyncConfiguration.java
@@ -40,8 +40,6 @@ public class SnapSyncConfiguration {
 
   public static final Boolean DEFAULT_SNAP_SYNC_BFT_ENABLED = Boolean.FALSE;
 
-  public static final Boolean DEFAULT_SNAP_SYNC_TO_HEAD_ENABLED_FLAG = Boolean.FALSE;
-
   public static SnapSyncConfiguration getDefault() {
     return ImmutableSnapSyncConfiguration.builder().build();
   }
@@ -89,10 +87,5 @@ public class SnapSyncConfiguration {
   @Value.Default
   public Boolean isSnapSyncBftEnabled() {
     return DEFAULT_SNAP_SYNC_BFT_ENABLED;
-  }
-
-  @Value.Default
-  public Boolean isSnapSyncToHeadEnabled() {
-    return DEFAULT_SNAP_SYNC_TO_HEAD_ENABLED_FLAG;
   }
 }
