+++ ethereum/eth/src/main/java/org/hyperledger/besu/ethereum/eth/sync/snapsync/SnapSyncMetricsManager.java
@@ -218,13 +218,11 @@ public class SnapSyncMetricsManager {
 
   public void notifySnapSyncCompleted() {
     final Duration duration = Duration.ofMillis(System.currentTimeMillis() - startSyncTime);
-    final long hours = (duration.toDaysPart() * 24) + duration.toHoursPart();
-
     LOG.info(
         "Finished worldstate snapsync with nodes {} (healed={}) duration {}{}:{},{}.",
         nbTrieNodesGenerated.addAndGet(nbTrieNodesHealed.get()),
         nbTrieNodesHealed,
-        hours > 0 ? (hours + ":") : "",
+        duration.toHoursPart() > 0 ? (duration.toHoursPart() + ":") : "",
         duration.toMinutesPart(),
         duration.toSecondsPart(),
         duration.toMillisPart());
